Patch v0.9.1

+++ grayc/src/runtime/array.h
@@ -32,17 +32,17 @@ GrayArray gray_array_new(GrayArena *arena, int32_t elem_size, int32_t initial_ca
 GrayArray gray_array_from(GrayArena *arena, const void *data, int32_t elem_size, int32_t count);
 
 /* Get a pointer to element at index (with bounds checking) */
-void *gray_array_get_ptr(GrayArray *arr, int32_t index, const char *file, int line);
+void *gray_array_get_ptr(GrayArray *arr, int64_t index, const char *file, int line);
 
 /* Set element at index (with bounds checking) */
-void gray_array_set(GrayArray *arr, int32_t index, const void *value, const char *file, int line);
+void gray_array_set(GrayArray *arr, int64_t index, const void *value, const char *file, int line);
 
 /* Ensure room for one more element, growing the backing store if full.
  * Growth is allocate-and-copy: the arena has no realloc, so the old store
  * lives on until the arena is reset or destroyed. Grows into the array's
  * owning arena when it has one, so an array reached through a mutable
  * reference does not end up backed by a shorter-lived scope arena.
- * file/line locate the P0035 panic on capacity overflow; pass NULL/0 for
+ * file/line locate the P0130 panic on capacity overflow; pass NULL/0 for
  * a panic without a source location. */
 void gray_array_grow(GrayArena *arena, GrayArray *arr, const char *file, int line);
